+++ src/Catalog/Repositories/MysqlCatalogRepositories.php
@@ -10,6 +10,7 @@ use Thinktomorrow\Trader\Application\Product\ProductDetail\ProductDetailReposito
 use Thinktomorrow\Trader\Application\Product\VariantLinks\VariantLinksComposer;
 use Thinktomorrow\Trader\Application\Taxon\Queries\TaxaSelectOptions;
 use Thinktomorrow\Trader\Application\Taxon\Queries\TaxonFilters;
+use Thinktomorrow\Trader\Application\Taxon\Queries\TaxonHierarchy;
 use Thinktomorrow\Trader\Application\Taxon\Redirect\TaxonRedirectRepository;
 use Thinktomorrow\Trader\Application\Taxon\Tree\TaxonTreeRepository;
 use Thinktomorrow\Trader\Domain\Model\Product\ProductRepository;
@@ -31,6 +32,7 @@ use Thinktomorrow\Trader\Infrastructure\Test\TestTraderConfig;
 use Thinktomorrow\Trader\Infrastructure\Vine\VineFlattenedTaxonIds;
 use Thinktomorrow\Trader\Infrastructure\Vine\VineTaxaSelectOptions;
 use Thinktomorrow\Trader\Infrastructure\Vine\VineTaxonFilters;
+use Thinktomorrow\Trader\Infrastructure\Vine\VineTaxonHierarchy;
 use Thinktomorrow\Trader\TraderConfig;
 
 class MysqlCatalogRepositories implements CatalogRepositories
@@ -62,12 +64,19 @@ class MysqlCatalogRepositories implements CatalogRepositories
             ->withMemoization(false);
     }
 
+    public function taxonHierarchy(): TaxonHierarchy
+    {
+        return new VineTaxonHierarchy($this->taxonTreeRepository());
+    }
+
     public function gridRepository(): GridRepository
     {
+        $taxonTreeRepository = $this->taxonTreeRepository();
+
         return new MysqlGridRepository(
             $this->container,
             $this->config,
-            new VineFlattenedTaxonIds($this->taxonTreeRepository())
+            new VineFlattenedTaxonIds($taxonTreeRepository, new VineTaxonHierarchy($taxonTreeRepository))
         );
     }
 
@@ -106,12 +115,12 @@ class MysqlCatalogRepositories implements CatalogRepositories
 
     public function taxonFilters(): TaxonFilters
     {
-        return new VineTaxonFilters(new TestTraderConfig, $this->taxonTreeRepository(), $this->taxonomyRepository());
+        return new VineTaxonFilters(new TestTraderConfig, $this->taxonTreeRepository(), $this->taxonomyRepository(), $this->taxonHierarchy());
     }
 
     public function flattenedTaxonIds(): FlattenedTaxonIds
     {
-        return new VineFlattenedTaxonIds($this->taxonTreeRepository());
+        return new VineFlattenedTaxonIds($this->taxonTreeRepository(), $this->taxonHierarchy());
     }
 
     public function taxaSelectOptions(): TaxaSelectOptions