+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Important changes will be notified in this file
 
 ## unreleased
 
+- Added: `TaxonHierarchy` access and shipping profile eligibility wiring for Trader 0.9.7.
+
 ## 2026-04-14 - 0.1.0
 
 - First release of testing package to be used on thinktomorrow/trader:0.9.*
-